+++ plugins/voxelman/core/chunkgen.d
@@ -20,6 +20,7 @@ import voxelman.core.config;
 import voxelman.storage.chunk;
 import voxelman.storage.chunkprovider;
 import voxelman.storage.coordinates;
+import voxelman.utils.compression;
 import core.thread;
 
 
@@ -47,10 +48,12 @@ struct LoadSnapshotMessage {
 }
 
 version = DBG_OUT;
+//version = DBG_COMPR;
 void chunkGenWorkerThread(shared(Worker)* workerInfo)
 {
 	import std.array : uninitializedArray;
 
+	ubyte[] compressBuffer = uninitializedArray!(ubyte[])(CHUNK_SIZE_CUBE*BlockId.sizeof);
 	try
 	{
 		void genChunk()
@@ -63,7 +66,7 @@ void chunkGenWorkerThread(shared(Worker)* workerInfo)
 			generator.genPerChunkData();
 
 			bool uniform = true;
-			BlockId[] blocks = uninitializedArray!(BlockId[])(CHUNK_SIZE_CUBE);
+			BlockId[CHUNK_SIZE_CUBE] blocks;
 
 			blocks[0] = generator.generateBlock(0, 0, 0);
 			BlockId uniformBlockId = blocks[0];
@@ -87,18 +90,45 @@ void chunkGenWorkerThread(shared(Worker)* workerInfo)
 			enum layerId = 0;
 			enum timestamp = 0;
 			enum numLayers = 1;
-			enum saved = false;
 
 			workerInfo.resultQueue.startPush();
-			workerInfo.resultQueue.pushItem(ChunkHeaderItem(cwp, numLayers, saved));
+			workerInfo.resultQueue.pushItem(ChunkHeaderItem(cwp, numLayers));
 			if(uniform)
+			{
 				workerInfo.resultQueue.pushItem(ChunkLayerItem(StorageType.uniform, layerId, 0, timestamp, uniformBlockId));
+			}
 			else
 			{
-				ushort dataLength = cast(ushort)blocks.length;
-				assert(dataLength == CHUNK_SIZE_CUBE);
-				ubyte* data = cast(ubyte*)blocks.ptr;
-				workerInfo.resultQueue.pushItem(ChunkLayerItem(StorageType.fullArray, layerId, dataLength, timestamp, data));
+				import core.memory : GC;
+				//infof("%s L %s B (%(%02x%))", cwp, blocks.length, cast(ubyte[])blocks);
+				ubyte[] compactBlocks = compress(cast(ubyte[])blocks, compressBuffer);
+				//infof("%s L %s C (%(%02x%))", cwp, compactBlocks.length, cast(ubyte[])compactBlocks);
+
+				if (compactBlocks.length <= ushort.max)
+				{
+					version(DBG_COMPR)infof("Gen1 %s %s %s\n(%(%02x%))", cwp, compactBlocks.ptr, compactBlocks.length, cast(ubyte[])compactBlocks);
+					compactBlocks = compactBlocks.dup;
+					version(DBG_COMPR)infof("Gen2 %s %s %s\n(%(%02x%))", cwp, compactBlocks.ptr, compactBlocks.length, cast(ubyte[])compactBlocks);
+					ushort dataLength = cast(ushort)compactBlocks.length;
+					ubyte* data = cast(ubyte*)compactBlocks.ptr;
+					// Add root to data.
+					// Data can be collected by GC if no-one is referencing it.
+					// It is needed to pass array trough shared queue.
+					GC.addRoot(data); // TODO remove when moved to non-GC allocator
+					workerInfo.resultQueue.pushItem(ChunkLayerItem(StorageType.compressedArray, layerId, dataLength, timestamp, data));
+				}
+				else
+				{
+					infof("Gen non-compressed %s", cwp);
+					ushort dataLength = cast(ushort)blocks.length;
+					assert(dataLength == CHUNK_SIZE_CUBE);
+					ubyte* data = cast(ubyte*)blocks.dup.ptr;
+					// Add root to data.
+					// Data can be collected by GC if no-one is referencing it.
+					// It is needed to pass array trough shared queue.
+					GC.addRoot(data); // TODO remove when moved to non-GC allocator
+					workerInfo.resultQueue.pushItem(ChunkLayerItem(StorageType.fullArray, layerId, dataLength, timestamp, data));
+				}
 			}
 			workerInfo.resultQueue.endPush();
 		}